Updated ConvASM3x3 solver test, rm FAIL conflict with GoogleTest
