+++ test/test.hpp
@@ -93,7 +93,13 @@ inline void expect_op(const TLeft& left,
 #define EXPECT_EQUAL(LEFT, RIGHT) EXPECT_OP(LEFT, ==, RIGHT)
 #define STATUS(...) EXPECT((__VA_ARGS__) == 0)
 
+// Conflicts with GoogleTest FAIL()
+#ifdef FAIL
+#undef FAIL
 #define FAIL(...) failed(__VA_ARGS__, __FILE__, __LINE__)
+#else
+#define FAIL(...) failed(__VA_ARGS__, __FILE__, __LINE__)
+#endif
 
 template <class F>
 bool throws(F f)
